core: vm: sqlvm: introduce go-sumtype
We have many interfaces decalred as a part of AST. Most of them are used
in a way called 'sum type', a type which is a collection of other types.
Since Go does not have a builtin way to check whether a type switch is
exhaustive, we add a static analysis tool called 'go-sumtype' to the CI
process to ensure we never hit the default panic case at runtime.

Since each interface must have at least a unique unexported method to
be considered as a valid sum type, we need a naming convention here.
Currently a unique name is created by putting ˉ character in front of
the name of the interface.

Remove Value function from Valuer interface declaration because the AST
printer has switched to reflection instead of using this interface. It
is not expected to have new users of Value functions because it returns
an empty interface. Type assertion is still required to access the value
and it is just clearer to assert the type of the AST node.

Remove AnyValueNode and DefaultValueNode from ExprNode. It is invalid to
use them as a part of an expression, and putting them in ExprNode just
makes type switches unreliable. Data types allocated for them are also
removed because non-expression nodes don't need data types.

Introduce InsertOptionNode and ColumnConstraintNode interfaces as sum
types. We used to use the generic Node interface to store them. Create
specific sum types enables go-sumtype to check type switches statically.

Drop assertNodeSlice because all callers are gone.